% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{categories}
\alias{categories}
\title{Get Coinmarketcap Categories}
\usage{
categories(ids, id_type = "id", exportfile = NULL)
}
\arguments{
\item{ids}{vector of ids}

\item{id_type}{one of 'id', 'slug' or 'symbol' depending on which coin idendifier you want to filter your coins by.}

\item{exportfile}{file path of where to save the data.}
}
\value{
A data frame containing the ID information for all cryptocurrencies available.
}
\description{
Uses the \emph{categories} API that returns the coin categories available in Coinmarketcap.
For more details see the \href{https://coinmarketcap.com/api/documentation/v1/#operation/getV1CryptocurrencyCategories}{documentation}.
}
\details{
You should use this function once at the beginning, save the file somewhere and then never
re-run this function again, to save on API calls.
}
