% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins_map.R
\name{coins_map}
\alias{coins_map}
\title{Get Coinmarketcap Coin Map}
\usage{
coins_map(
  listing_status = c("active", "inactive", "untracked"),
  aux = c("platform", "first_historical_data", "last_historical_data",
    "is_active,status"),
  exportfile = NULL
)
}
\arguments{
\item{listing_status}{Character vector containing the type of currencies you want to include. Can be "active", "inactive" or
"untracked". Defaults to all the three.}

\item{aux}{Character vector with auxiliary information. Defaults to all possible info.}

\item{exportfile}{If output path is included, will save data to that file.}
}
\value{
A data frame containing the ID information for all cryptocurrencies available.
}
\description{
Uses the \emph{map} API that returns header information from Coinmarketcap.
For more details see the \href{https://coinmarketcap.com/api/documentation/v1/#operation/getV1CryptocurrencyMap}{documentation}.
}
\details{
You should use this function once at the beginning, save the file somewhere and then never
re-run this function again, to save on API calls.
}
