% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Get Coinmarketcap Coin Metadata.}
\usage{
metadata(type = "id", identifiers, exportfile = NULL)
}
\arguments{
\item{type}{Type of identifiers used to search the coins. Can be one of 'id', 'slug', 'symbol' or 'address'. Defaults to Coinmarketcap id.}

\item{identifiers}{Vector containing the list of currencies you want to include.}

\item{exportfile}{If output path is included, will save data to that file.}
}
\value{
A data frame containing the metadata information for all cryptocurrencies available.
}
\description{
Uses the \emph{info} API that returns header information from Coinmarketcap.
For more details see the \href{https://coinmarketcap.com/api/documentation/v1/#operation/getV2CryptocurrencyInfo}{documentation}.
}
\details{
You should use this function once at the beginning, save the file somewhere and then never
re-run this function again, to save on API calls.
}
